corrected error in send_recv_atoms_data call: passing int instead of double

+++ hdr/dipole.hpp
@@ -86,7 +86,7 @@ namespace dipole{
                            std::vector<int>& cells_num_atoms_in_cell,
                            int cells_num_local_cells,
                            int cells_num_cells,
-                           int cells_macro_cell_size
+                           double cells_macro_cell_size
                            );
    /*------------------------------------------------*/
    /*Function to sort cells/atoms data after sharing */
